requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.1.1
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
